Improved configuration of some objects, added benches and many more enemies

+++ WorldEditor/Architect/Objects/SelectableObject.cs
@@ -14,7 +14,7 @@ public abstract class SelectableObject
 
     public abstract int GetWeight();
 
-    public virtual int GetSpriteRotation()
+    public virtual float GetSpriteRotation()
     {
         return 0;
     }
